move hardware to new module
