+++ hcam_drivers/hardware/__init__.py
@@ -1,0 +1,10 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+# This sub-module is destined for common non-package specific utility
+# functions that will ultimately be merged into `astropy.utils`
+
+# As far as possible, utils contains classes of generic use, such as
+# PosInt for positive integer input. See e.g. 'hcam' for more instrument
+# dependent components.
+
+from __future__ import print_function, unicode_literals, absolute_import, division
