+++ hcam_drivers/hardware/honeywell.py
@@ -1,6 +1,6 @@
 # talk to honeywell temperature monitor
 from __future__ import absolute_import, unicode_literals, print_function, division
-from .utils import DriverError
+from ..utils import DriverError
 import six
 if not six.PY3:
     from pymodbus.constants import Endian
